+++ src/components/Header.tsx
@@ -1,6 +1,8 @@
 import { useSession, signIn } from "next-auth/react";
 import Router, { useRouter } from "next/router";
 import Image from "next/future/image";
+import { useAtom } from "jotai";
+import { showFilterAtom, filterQueryAtom } from "../store";
 import { UserMenu } from "./UserMenu";
 import { getNow, flattenQuery } from "../../shared/utils";
 import Logo from "../../public/static/images/bath-cc-logo.svg";
@@ -23,14 +25,20 @@ const LinkButton = ({ children, ...props }: ButtonProps) => (
 const goHome = () => Router.push("/");
 
 export const Header = () => {
+  const [, setShowFilterMenu] = useAtom(showFilterAtom);
+  const [filterQuery] = useAtom(filterQueryAtom);
   const { status, data: session } = useSession();
   const isAuthenticated = status === "authenticated";
   const role = session?.role as string;
   const router = useRouter();
   const rideId = router.query.id;
   const rideDate = router.query.date;
-
+  const isRidesPage = router.pathname === "/";
   const isHistoric = flattenQuery(rideDate) < getNow();
+
+  const showFilters = () => setShowFilterMenu(true);
+
+  const hasFiltersApplied = !!(filterQuery.onlyJoined || filterQuery.q);
 
   return (
     <div className="fixed z-10 flex h-16 w-full items-center justify-center bg-blue-900  text-white sm:h-24 md:bg-slate-100 md:text-neutral-700">
@@ -50,7 +58,23 @@ export const Header = () => {
             BCC Rides
           </button>
         </div>
-        <div className="flex items-center">
+
+        <div className="flex items-center gap-4">
+          {isRidesPage && (
+            <button
+              type="button"
+              onClick={showFilters}
+              title="Filter results"
+              className="flex items-center rounded p-1 text-3xl md:hover:bg-slate-200"
+            >
+              {hasFiltersApplied ? (
+                <i className="fa-solid fa-filter-circle-xmark" />
+              ) : (
+                <i className="fa-solid fa-filter" />
+              )}
+            </button>
+          )}
+
           {isAuthenticated ? (
             <UserMenu role={role} rideId={rideId} isHistoric={isHistoric} />
           ) : (