+++ src/components/index.tsx
@@ -6,6 +6,7 @@ export * from "./Button/JoinButton";
 export * from "./Calendar";
 export * from "./Card";
 export * from "./Confirm";
+export * from "./Filters";
 export * from "./forms/AnonymousUserForm";
 export * from "./forms/RideForm";
 export * from "./forms/UserProfileForm";