+++ src/types/index.ts
@@ -68,3 +68,9 @@ export type SeasonStartTime = {
   winter: StartTime;
   summer: StartTime;
 };
+
+export type FilterQuery = {
+  onlyJoined?: boolean;
+  q?: string;
+  weeksAhead?: string;
+};